requirements.txt
python-decouple
pandas
requests
tqdm